Rename Mechadendrite Slots
Reason for change: Mechadendrite spam causes the surgery screen to be hard to read. First, second, third, etc are extremely hard to differentiate when it's so crowded. Mechadendrites also do not organize neatly due to a vanilla issue so ordinal numbers do not work well. Hope is cardinal numbers are easier to read and distinguish.